fix: 26 bugs + UX improvements + performance optimizations
Bug fixes:
- Remove dead getLatestUpload() calling non-existent endpoint
- Settings password: add uppercase/lowercase/digit validation
- Upload buildParsedData: fix float drift with integer-cents math
- handleRestore: simplified to metadata-only, no full record load
- finalizeImport: navigation only on success, not on error

UX improvements:
- SessionRestorer: rewritten to use listUploads metadata only
- useUnifiedAuditLog: add staleTime 2min
- AuthContext: prefetchQuery uploads after login
- Transactions: add search spinner during 400ms debounce
- Dashboard: add Upload File CTA on empty state with Next.js Link

Sidebar redesign:
- Replace hover-expand with toggle arrow button on right edge
- Auto-collapse on nav link click
- Collapse sidebar on click outside (invisible overlay)
- Collapse state wired through SidebarContext

Performance:
- useAuditLog: add staleTime 5min + gcTime 15min
- DynamicChart: move tooltip styles to module-level constants
- DynamicChart: wrap formatters in useCallback

+++ backend/app/models/audit_log.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, Float, ForeignKey, Integer, String, Text, func
+from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, Numeric, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
@@ -24,7 +24,7 @@ class AuditLog(Base):
     file_name: Mapped[str] = mapped_column(String(255), nullable=False)
     session_id: Mapped[str | None] = mapped_column(String(36), nullable=True)
     record_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    total_amount: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
+    total_amount: Mapped[float] = mapped_column(Numeric(15, 2), nullable=False, default=0.0)
     details: Mapped[str | None] = mapped_column(String(500), nullable=True)
     snapshot_data: Mapped[str | None] = mapped_column(Text, nullable=True)
     is_undone: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
